opvap

+++ src/pages/Fruits/Fruits.js
@@ -8,22 +8,24 @@ const Fruits = () => {
     const [fruits, setFruits] = useFruits();
 
     const handleDelete = (id) => {
-        // console.log('delete', id);
-        const url = `https://shielded-oasis-06280.herokuapp.com/fruits/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                // console.log(data);
-                const rem = fruits.filter(fruit => fruit._id !== id);
-                setFruits(rem);
+        const proceed = window.confirm("আপনি কী এই পণ্যটি ডিলিট করতে চাচ্ছেন?");
+        if (proceed) {
+            // console.log('delete', id);
+            const url = `https://shielded-oasis-06280.herokuapp.com/fruits/${id}`;
+            fetch(url, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    // console.log(data);
+                    const rem = fruits.filter(fruit => fruit._id !== id);
+                    setFruits(rem);
+                })
+        }
     }
 
     return (
         <div className='mt-5'>
-            <h2>FRUITS</h2>
             <Row xs={1} md={3} className="w-75 my-4 mx-auto g-4">
                 {
                     fruits.map(fruit => <Col key={fruit._id}>
@@ -32,7 +34,6 @@ const Fruits = () => {
                             <Card.Body>
                                 <Card.Title>নাম: {fruit.name}</Card.Title>
                                 <Card.Text>দাম: {fruit.price}টাকা/মণ</Card.Text>
-                                <Card.Text>পরিমাণ: {fruit.quantity}টন</Card.Text>
                                 <Card.Text>সরবরাহকারী: {fruit.supplierName}</Card.Text>
                             </Card.Body>
                             <Card.Footer>
@@ -46,7 +47,7 @@ const Fruits = () => {
                 }
             </Row>
             <div style={{ margin: '150px auto 40px auto' }} className='w-100'>
-                <Link className='btn btn-primary fs-3 w-50' to="/addfruits">Add New Fruits</Link>
+                <Link className='btn btn-primary fs-3 w-50' to="/addfruits">নতুন পণ্য যোগ করুন</Link>
             </div>
         </div>
     );
